Changed data type of index to avoid warnings

+++ main.cpp
@@ -9,7 +9,7 @@ int main()
     srand(time(0));
 
     int size = 32, array[size];
-    for (size_t i = 0; i < size; i++)
+    for (int i = 0; i < size; i++)
     {
         array[i] = rand() % 64;
     }
